\name{demo.abi.multiple}
\alias{demo.abi.multiple}
\docType{data}
\title{Demonstration vector referring to an ABI output file with multiple profiles}
\description{
  A file that contains a collection of profiles for demonstrating the RiboSort package. It is a subset of a full samples database, and is not intended to represent any real data set.
}
\usage{data(demo.abi.multiple)}
\format{A character element: "abimultiple.txt"}
\examples{
data(demo.abi.multiple)
demo.abi.multiple
}
\keyword{datasets}
