\name{demo.abi.single}
\alias{demo.abi.single}
\docType{data}
\title{Demonstration vector referring to an ABI output file with a single profile}
\description{
  A vector listing three filenames, each of which corresponds to a single community profile produced by an ABI sequencer. These files are used for demonstrating the RiboSort package. They are a subset of a full samples database, and is not intended to represent any real data set.
}
\usage{data(demo.abi.single)}
\format{
  A character vector of length three. 
}

\examples{
data(demo.abi.single)
demo.abi.single
}
\keyword{datasets}
