\name{demo.beckman}
\alias{demo.beckman}
\docType{data}
\title{Demonstration list of profiles produced by the Beckman Sequencer}
\description{
  A vector listing four filenames, each of which corresponds to a single community profile produced by the Beckman Coulter sequencer. These files are used for demonstrating the RiboSort package. They are a subset of a full samples database, and is not intended to represent any real data set.
}
\usage{data(demo.beckman)}
\format{
  A character vector of length four. 
}

\examples{
data(demo.beckman)
demo.beckman
}
\keyword{datasets}
