\name{demo.standard}
\alias{demo.standard}
\docType{data}
\title{Demonstration list of profiles in Standard Profile form}
\description{
  A vector listing six filenames, each of which corresponds to a single community profile in standard profile form. These files are used for demonstrating the RiboSort package. They are a subset of a full samples database, and is not intended to represent any real data set.
}
\usage{data(demo.standard)}
\format{
  A character vector of length six.
}
\examples{
data(demo.standard)
demo.standard
}
\keyword{datasets}
