\name{RiboSort-package}
\alias{RiboSort-package}
\docType{package}
\title{
Data Preparation and Analysis of Microbial Community Profiles
}
\description{
RiboSort is an R package for automated data preparation and exploratory analysis of (TRFLP & ARISA) microbial community profiles. Its function is to eliminate the laborious task of manually classifying community fingerprints in microbial studies. By automatically assigning detected fragments and their respective relative abundances to appropriate ribotypes, RiboSort saves time and greatly simplifies the preparation of DNA fingerprint datasets for statistical analysis. This package also provides a platform for effortless creation of multi-dimensional scaling plots of the data sorted by RiboSort.
}
\details{
For a detailed description of RiboSorts functionality and how to use it, consult the RiboSort package vignette.
}
\author{
Una Scallan
Maintainer: Una Scallan <una.scallan@ucd.ie>
}
\keyword{manip}
\keyword{package}
\keyword{multivariate}

